Moved utils into its own module
