+++ app.js
@@ -1,6 +1,12 @@
-require("./app.css");
-const io = require("./dependencies/socket.io.js");
-const SimpleSignalClient = require("./dependencies/simple-signal-client.min.js");
+// Modules
+import utils from "./modules/utils"
+
+// App statics
+require("./app.css")
+
+// Dependencies
+const io = require("./dependencies/socket.io")
+const SimpleSignalClient = require("./dependencies/simple-signal-client.min")
 
 var w = window,
 d = document,
@@ -18,19 +24,7 @@ canvas.setAttribute('height', screenHeight);
 var ctx = canvas.getContext("2d");
 ctx.strokeStyle = "#FF0000";
 
-// todo: move to another module (duh)
-var util = {};
-util.getParameterByName = function (name) {
-  var url = window.location.href
-  name = name.replace(/[[\]]/g, '\\$&')
-  var regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)')
-  var results = regex.exec(url)
-  if (!results) return null
-  if (!results[2]) return ''
-  return decodeURIComponent(results[2].replace(/\+/g, ' '))
-}
-
-var room = util.getParameterByName('room') || '1'
+var room = utils.getParameterByName('room') || '1'
 
 var socket = io('https://kaleidraw-signal-server-bglcynwuea.now.sh');
 var signal = new SimpleSignalClient(socket, {room: room});
