+++ modules/utils.js
@@ -1,0 +1,13 @@
+const utils = {};
+
+utils.getParameterByName = (name) => {
+  var url = window.location.href
+  name = name.replace(/[[\]]/g, '\\$&')
+  var regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)')
+  var results = regex.exec(url)
+  if (!results) return null
+  if (!results[2]) return ''
+  return decodeURIComponent(results[2].replace(/\+/g, ' '))
+};
+
+module.exports = utils;
